Separating in classes
