+++ Net2Plan-GUI/Net2Plan-GUI-Plugins/Net2Plan-NetworkDesign/src/main/java/com/net2plan/gui/plugins/networkDesign/viewEditTopolTables/controlTables/LastRowAggregatedValue.java
@@ -1,0 +1,31 @@
+package com.net2plan.gui.plugins.networkDesign.viewEditTopolTables.controlTables;
+
+public class LastRowAggregatedValue
+{
+    private String value;
+
+    public LastRowAggregatedValue()
+    {
+        value = "---";
+    }
+
+    public LastRowAggregatedValue(int val)
+    {
+        value = "" + val;
+    }
+
+    public LastRowAggregatedValue(double val)
+    {
+        value = String.format("%.2f", val);
+    }
+
+    public LastRowAggregatedValue(String value)
+    {
+        this.value = value;
+    }
+
+    public String toString()
+    {
+        return value;
+    }
+}
